wip simplifying the whole gobot logic. debugging.
